Carry full 12-bit resolution end to end, not just at the PWM stage
The previous PWM-resolution fix only widened the gamma LUT's output; the
"brightness" domain it was indexed by was still a uint8_t (0-255), even
though the pot's ADC reads 12-bit and gamma correction is computed in
float. That meant every pot reading was crushed from 4096 possible
positions down to 256 discrete brightness levels before gamma or PWM
ever saw it, capping the lamp to 256 truly distinguishable output levels
regardless of downstream resolution.

MAX_BRIGHTNESS is now 4095, deliberately equal to both POT_ADC_MAX and
led_control.cpp's PWM_MAX_DUTY, so mapPotToBrightness() becomes an
identity mapping and the pot's full ADC resolution passes through
untouched: ADC(12-bit) -> brightness(12-bit) -> gamma LUT(4096 entries)
-> PWM(12-bit), with no intermediate rounding anywhere in the chain.

Widened alongside brightness itself: the gamma LUT and its indexing,
the brightness slew target, RTC-persisted per-mode brightness (button
mode), and getBatteryLimitedMaxBrightness(). All brightness-relative
config thresholds (POT_OFF_THRESHOLD, POT_ON_HYSTERESIS,
POT_MAX_DEADZONE, POT_MOVEMENT_DEADBAND, CRITICAL_MAX_BRIGHTNESS,
PULSE_MIN_BRIGHTNESS, MIN_BRIGHTNESS_PWM) are rescaled by the same
factor so their meaning (percentages, floors) is unchanged. Button
mode's long-press step size is likewise scaled (new
BRIGHTNESS_STEP_SIZE) so discrete dimming doesn't get 16x slower for
the same hold duration — only the pot's live tracking needed the wider
domain.

+++ firmware/src/battery_monitor.cpp
@@ -61,7 +61,7 @@ float getLastBatteryVoltage() {
   return lastBatteryVoltage;
 }
 
-uint8_t getBatteryLimitedMaxBrightness() {
+uint16_t getBatteryLimitedMaxBrightness() {
   if (currentBatteryState == BATTERY_CRITICAL) {
     return CRITICAL_MAX_BRIGHTNESS;
   }
